Added Pagination in Competition Table

+++ src/app/competition/competition.component.ts
@@ -1,35 +1,45 @@
 import { Component, AfterViewInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { CompetitionService } from '../service/competition.service';
-//declare var require: any;
+import { PageEvent } from '@angular/material/paginator';
+import Swal from 'sweetalert2';
 
 @Component({
   templateUrl: './competition.component.html'
 })
-export class CompetitionComponent implements AfterViewInit {
+export class CompetitionComponent {
   
-  competitions!:any[];
+  competitions: any[] = [];
   selectedStatus: string = '';
+  pageSizeOptions: number[] = [5, 10, 20];
+  pageSize: number = 5;
+  pageIndex: number = 0;
+  totalCompetitions: number = 0;
 
-  constructor(private competitionService:CompetitionService, private router:Router) {
-    
-  }
-
-  ngAfterViewInit() { }
+  constructor(private competitionService: CompetitionService, private router: Router) {}
 
   ngOnInit() {
-   this.getCompetitions()
+    this.getCompetitions();
   }
 
   getCompetitions() {
-    this.competitionService.getCompetitions().subscribe(
-      (response) => {
-        this.competitions = response.data;
-      },
-      (error) => {
-        console.error('Error fetching Competitions:', error);
-      }
-    );
+    this.competitionService
+      .getCompetitions(this.pageIndex, this.pageSize)
+      .subscribe(
+        (response) => {
+          this.competitions = response.data;
+          this.totalCompetitions = response.data.length;
+        },
+        (error) => {
+          console.error('Error fetching Competitions:', error);
+        }
+      );
+  }
+
+  onPageChange(event: PageEvent) {
+    this.pageIndex = event.pageIndex;
+    this.pageSize = event.pageSize;
+    this.getCompetitions();
   }
 
   getCompetitionsByStatus() {
@@ -52,14 +62,38 @@ export class CompetitionComponent implements AfterViewInit {
   }
 
   deleteCompetition(id: number) {
-    this.competitionService.deleteCompetition(id).subscribe(
-      (response) => {
-        console.log('Competition deleted successfully');
-      },
-      (error) => {
-        console.error('Error deleting competition', error);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this competition!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.competitionService.deleteCompetition(id).subscribe(
+          (response) => {
+            Swal.fire({
+              icon: 'success',
+              title: 'Competition Deleted Successfully',
+              showConfirmButton: false,
+              timer: 1500
+            });
+            this.getCompetitions();
+          },
+          (error) => {
+            Swal.fire({
+              icon: 'error',
+              title: 'Error Deleting Competition',
+              text: 'An error occurred while deleting the competition. Please try again.',
+              confirmButtonColor: '#d33',
+              confirmButtonText: 'OK'
+            });
+          }
+        );
       }
-    );
+    });
   }
 
 }
